refactor: 5단계 리뷰 반영
Simulator 삭제
Car의 record 삭제
strategy package 분류
step5, step2 패키지명 변경
static 생성자에서 parameter 검증
에러 클래스 추가
