+++ src/main/java/racingcar/domain/exception/InvalidCarCountException.java
@@ -1,0 +1,7 @@
+package racingcar.domain.exception;
+
+public class InvalidCarCountException extends RuntimeException {
+    public InvalidCarCountException() {
+        super("invalid car count");
+    }
+}
