+++ src/main/java/racingcar/domain/exception/InvalidCarNameLengthException.java
@@ -1,0 +1,7 @@
+package racingcar.domain.exception;
+
+public class InvalidCarNameLengthException extends RuntimeException {
+    public InvalidCarNameLengthException() {
+        super("invalid car name length");
+    }
+}
