+++ src/main/java/racingcar/domain/exception/InvalidTryCountException.java
@@ -1,0 +1,7 @@
+package racingcar.domain.exception;
+
+public class InvalidTryCountException extends RuntimeException {
+    public InvalidTryCountException() {
+        super("invalid try count");
+    }
+}
