+++ src/main/java/racingcar/domain/strategy/RacingStrategy.java
@@ -1,0 +1,7 @@
+package racingcar.domain.strategy;
+
+import racingcar.domain.car.Car;
+
+public interface RacingStrategy {
+    void move(Car car);
+}
