+++ src/main/java/racingcar/domain/strategy/RandomRacingStrategy.java
@@ -1,0 +1,20 @@
+package racingcar.domain.strategy;
+
+import racingcar.domain.car.Car;
+import racingcar.util.RandomGenerator;
+
+public class RandomRacingStrategy implements RacingStrategy {
+    public RandomRacingStrategy() {
+    }
+
+    @Override
+    public void move(Car car) {
+        if (shouldMove()) {
+            car.move();
+        }
+    }
+
+    private boolean shouldMove() {
+        return ((int) (RandomGenerator.getRandomLessThan(10))) >= 4;
+    }
+}
