+++ src/main/java/racingcar/util/RandomGenerator.java
@@ -1,0 +1,10 @@
+package racingcar.util;
+
+public class RandomGenerator {
+    private RandomGenerator() {
+    }
+
+    public static double getRandomLessThan(int number) {
+        return Math.random() * number;
+    }
+}
