--- src/main/java/step4/domain/RacingStrategy.java
@@ -1,5 +1,0 @@
-package step4.domain;
-
-public interface RacingStrategy {
-    void move(Car car);
-}
