--- src/main/java/step4/util/RandomGenerator.java
@@ -1,10 +1,0 @@
-package step4.util;
-
-public class RandomGenerator {
-    private RandomGenerator() {
-    }
-
-    public static double getRandomLessThan(int number) {
-        return Math.random() * number;
-    }
-}
